Access the EDGE skating speed statistics for a team, season, game type, and category

team_edge_skating_speed() scrapes the EDGE skating speed statistics for a given set of team, season, game_type, and category.

Usage

team_edge_skating_speed(
  team = 1,
  season = "now",
  game_type = "",
  category = "details"
)

Arguments

team

integer ID (e.g., 21), character full name (e.g., 'Colorado Avalanche'), OR three-letter code (e.g., 'COL'); see teams() for reference; ID is preferable as there now exists duplicate three-letter codes (i.e., 'UTA' for 'Utah Hockey Club' and 'Utah Mammoth')

season

integer in YYYYYYYY (e.g., 20242025); see team_edge_seasons() for reference

game_type

integer in 1:3 (where 1 = pre-season, 2 = regular season, 3 = playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'; see team_edge_seasons() for reference; most functions will NOT support pre-season

category

character of 'd'/'details' or 't'/'top'/'top speeds'

Value

data.frame with one row per position (category = 'details') or burst (category = 'top speeds')

Examples

COL_top_speeds_regular_20242025 <- team_edge_skating_speed(
  team      = 21,
  season    = 20242025,
  game_type = 2,
  category  = 'T'
)