Get goalie statistics

get_goalie_statistics() retrieves information on each goalie or game for a given set of season, teams, game_types, and report. dates must be given when paired with is_game as the default range will return incomplete data (too wide). Access get_configuration() for what information each combination of report, is_aggregate and is_game can provide. Access get_seasons() for season and dates and get_teams() for teams references. Will soon be reworked for easier access.

Usage

get_goalie_statistics(
  season = get_season_now()$seasonId,
  teams = 1:100,
  game_types = 1:3,
  dates = c("2025-01-01"),
  report = "summary",
  is_aggregate = FALSE,
  is_game = FALSE
)

Arguments

season

integer in YYYYYYYY

teams

vector of integers Team ID(s)

game_types

vector of integers where 1=pre-season, 2=regular, and 3=playoffs

dates

vector of strings in 'YYYY-MM-DD'

report

string

is_aggregate

boolean

is_game

boolean

Value

tibble with one row per goalie or game

Examples

playoff_goalie_svr_20242025 <- get_goalie_statistics(
  season=20242025,
  teams=1:100,
  game_types=c(3),
  report='startedVsRelieved'
)